REF: Merge Cloud Run and Dataflow answer_question modules
